+++ examples/multiFileProject/multiFileProject.h
@@ -1,0 +1,19 @@
+/****************************************************************************************************************************
+  multiFileProject.h
+  For Ethernet shields using WT32_ETH01 (ESP32 + LAN8720)
+
+  WebServer_WT32_ETH01 is a library for the Ethernet LAN8720 in WT32_ETH01 to run WebServer
+
+  Based on and modified from ESP8266 https://github.com/esp8266/Arduino/releases
+  Built by Khoi Hoang https://github.com/khoih-prog/WebServer_WT32_ETH01
+  Licensed under MIT license
+*****************************************************************************************************************************/
+
+// To demo how to include files in multi-file Projects
+
+#pragma once
+
+#define _ETHERNET_WEBSERVER_LOGLEVEL_       1
+
+// Can be included as many times as necessary, without `Multiple Definitions` Linker Error
+#include "WebServer_WT32_ETH01.hpp"