+++ src/WebServer_WT32_ETH01_Impl.h
@@ -12,7 +12,7 @@
   @file       Esp8266WebServer.h
   @author     Ivan Grokhotkov
   
-  Version: 1.4.1
+  Version: 1.5.0
 
   Version Modified By   Date      Comments
   ------- -----------  ---------- -----------
@@ -22,12 +22,15 @@
   1.3.0   K Hoang      23/10/2021 Making compatible with breaking core v2.0.0+
   1.4.0   K Hoang      27/11/2021 Auto detect ESP32 core version
   1.4.1   K Hoang      29/11/2021 Fix bug in examples to reduce connection time
+  1.5.0   K Hoang      21/06/2022 Fix multiple-definitions linker error
  *****************************************************************************************************************************/
 
 #pragma once
 
 #ifndef WEBSERVER_WT32_ETH01_IMPL_H
 #define WEBSERVER_WT32_ETH01_IMPL_H
+
+//////////////////////////////////////////////////////////////
 
 bool WT32_ETH01_eth_connected = false;