+++ changelog.md
@@ -12,6 +12,7 @@
 
 
 * [Changelog](#changelog)
+  * [Releases v1.5.0](#releases-v150)
   * [Releases v1.4.1](#releases-v141)
   * [Releases v1.4.0](#releases-v140)
   * [Releases v1.3.0](#releases-v130)
@@ -23,6 +24,11 @@
 ---
 
 ## Changelog
+
+#### Releases v1.5.0
+
+1. Fix multiple-definitions linker error.
+2. Add example [multiFileProject](examples/multiFileProject) to demo for multiple-file project to avoid `multiple-definitions` linker error
 
 #### Releases v1.4.1